requirements.txt
streamlit
google-genai
pillow
gspread
oauth2client
pymupdf
python-dotenv
